Fixes for WSL and consumer GPUs
Fix a crash in BCM switch code when PCI vendor is NULL.

Don't consider errors from NVML CC config calls to be fatal; they are
not supported on non-datacenter GPUs.

Don't try to unlink a previously unlinked shmem file.

BUG 4658355

+++ src/misc/shmutils.cc
@@ -158,7 +158,7 @@ ncclResult_t ncclShmUnlink(ncclShmHandle_t handle) {
   ncclResult_t ret = ncclSuccess;
   struct shmHandleInternal* tmphandle = (struct shmHandleInternal*)handle;
   if (tmphandle) {
-    if (tmphandle->shmPath != NULL) {
+    if (tmphandle->shmPath != NULL && tmphandle->refcount != NULL && *tmphandle->refcount > 0) {
       if (unlink(tmphandle->shmPath) != 0) {
         WARN("unlink shared memory %s failed, error: %s (%d)", tmphandle->shmPath, strerror(errno), errno);
         ret = ncclSystemError;
